Separated out ProductSearch class & tests.
